Move new rotate/flip/normalise functions to utils.js
